requirements.txt
django==5.2.4
python-decouple
psycopg[binary]
djangorestframework-gis
paho-mqtt
drf-yasg
haversine